feat: load user defined function during build
By `require(userFn)` during the build phase, we can detect invalid `package.json` and invalid `main: "yourfile.js"` values. This does mean we pay the cost of `require` of the user function and its dependencies during build. It also means that code outside of the function may execute during `require`. If, say, a user were to code in database migrations or something dangerous in their `index.js` that code would be attempted during the build's `require` instead of on function pod start. I'm not sure we have any guidelines past what the funciton should look like today.
